fix: removed browser mode toggle
