--- Assets/Packages.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 19f17ccaf65d34764afddc8321c71b84
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
