--- Assets/Resources.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bb6c9e99d2f25314f8609da6018623da
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
