--- Assets/Sprites.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4277298a7d0cf474ba470ba3a2c375e3
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
