--- Assets/StreamingAssets.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f8f3ec6b5ea99a541be2156397fd97af
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
