+++ ProjectSettings/EditorBuildSettings.asset
@@ -8,7 +8,7 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Scenes/Lobby.unity
     guid: 27215d4ccddeb6d49af03911dd97f9b1
-  - enabled: 1
+  - enabled: 0
     path: Assets/PlayroomKit/Examples/2d-platformer/2d-platformer.unity
     guid: cf82eec4c654cf44b89cec2a91afebd5
   - enabled: 0
@@ -17,4 +17,7 @@ EditorBuildSettings:
   - enabled: 0
     path: Assets/Scenes/topdown.unity
     guid: 490247da3cf48784ab657a183ba059d1
+  - enabled: 1
+    path: Assets/PlayroomKit/Examples/package-showcase/showcase.unity
+    guid: 85c9af77ae006408ab24500b953a9035
   m_configObjects: {}
